Get request wait time from server

+++ core/src/main/java/com/gigigo/orchextra/core/data/datasources/network/models/ApiConfiguration.kt
@@ -23,7 +23,7 @@ data class ApiConfiguration(
     val beaconRegions: List<ApiRegion>?,
     val eddystoneRegions: List<ApiRegion>?,
     val customFields: Map<String, ApiCustomField>?,
-    val requestWaitTime: Int?,
+    val requestWaitTime: Long?,
     val vuforia: ApiVuforia?)
 
 data class ApiGeofence(
